Add support for MCUboot
Using MCUboot with two fallbacks if the device ever gets stuck.
A: If there is no bootable application, the bootloader goes into
   USB DFU mode.
B: If the Button for USB-Port 6 is pressed during boot, the bootloader
   goes into USB DFU mode.
This way, there is always a way to upgrade the firmware on the STM if
anything goes wrong during firmware upgrade.